Retool packet boundaries to match SvtAv1EncApp
Packets were oriented around encoded pictures, whereas most stuff
seems to expect packets oriented around displayed pictures.
The standalone SvtAv1EncApp has logic to retool the packets,
which is adapted here.

pts is stored in a queue to match back up with the one-in-one-out
picture->packet expectation.

Also, AV1 extradata is forced on, which seems to be required to
get things working with the HLS muxer. Need to investigate this
further, as it looks like it ought to request it already.